requirements.txt
Pillow
pysmb
